pyscript/requirements.txt
flask
formsg-python-sdk
dotenv
python-dotenv
typing-extensions>=4.10.0
gunicorn